Crate soundcloud

SoundCloud API library

This soundcloud library provides an interface where you can query soundcloud for information about tracks and users.

Reexports

pub use error::Error;

Modules

error

Structs

App
Registered client application.

Client

Comment
User comment.

Track
Uploaded track.

User
Registered user.

Constants

API_HOST
The static host address for the API.